+++ src/Main.java
@@ -5,9 +5,9 @@ public class Main {
         Output output = new Output();
         Solution solution = new Solution();
 
-        int[] C = input.inputC();
-        int[][] A = input.inputA();
-        int[] B = input.inputB();
+        double[] C = input.inputC();
+        double[][] A = input.inputA();
+        double[] B = input.inputB();
         int approximation = input.inputApproximation();
         boolean min = input.min();
